pyproject.toml
[tool.poetry]
name = "hachathon"
version = "0.1.0"
description = ""
package-mode = false
authors = []


[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "*"
fsspec = "*"
s3fs = "*"
pandasql = "*"
streamlit = "*"
langchain_aws = "*"
snowflake-connector-python = "*"
pandas = "*"
ipykernel = "*"
langchain-community = "*"
langgraph = "*"
langgraph-checkpoint-sqlite = "*"
matplotlib = "^3.9.2"
langchain-experimental = "^0.3.2"
scikit-learn = "^1.5.2"
imblearn = "^0.0"
xgboost = "^2.1.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

